metrics: collectd: scale y axis to largest of either
With the previous `x= max(c(x, blah))`, and a default of `x=1` (for
cpu_scale for instance), the graph would only scale 'down to' one of
the y axis, which for some data sets (large pod counts for instance)
resulted in the CPU Y axis being set to '250' say, but the cpu data
topping out at 100 (%). This mean the CPU data was squashed into the
bottom 3rd of the Y axis.

Drop the `max(c(1,x))`, so we set the scaling factor from 0..x, rather
than 1..x, and thus fit both sets of data to fill the graph area.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>